Game Activity and basic deck of cards (next card) functionality
Creates a shuffled deck when clicking on Start Solo Deck. Pulls a card
from the deck when you click "done with this card". Displays debugging
info for remaining cards in deck, current card, etc. Created a
"finished with deck" card as the last card. Timing is next on the list.
Note that using the back button, and re-clicking on the Start Solo Deck
will create a new deck and ignore any previous state. Ideally Game
class can be used by both Solo and Coop modes.